Finished KdbxWriter work for KDBX4. FIxes #187.

+++ PassKeep.Lib/KeePass/Kdf/Argon2Parameters.cs
@@ -219,7 +219,7 @@ namespace PassKeep.Lib.KeePass.Kdf
             dict[SecretKey] = new VariantValue(this.secretKey);
             dict[DataKey] = new VariantValue(this.associatedData);
             dict[SaltKey] = new VariantValue(this.salt);
-            dict[VersionKey] = new VariantValue(Argon2d.VersionNumber);
+            dict[VersionKey] = new VariantValue((uint)Argon2d.VersionNumber);
 
             return dict;
         }
